Separated the device and environment metrics into their own files.
